Typescript

+++ src/components/incomes/incomes.ts
@@ -1,50 +1,56 @@
 import * as bootstrap from 'bootstrap';
 import {HttpUtils} from "../../utils/http-utils";
+import {OpenNewRouteType} from "../../types/openNewRoute.type";
+import {GetCategoryType} from "../../types/get-category.type";
 
 export class Incomes {
-    constructor(openNewRoute) {
-        this.openNewRoute = openNewRoute;
+    readonly openNewRoute: OpenNewRouteType;
 
+    constructor(openNewRoute: OpenNewRouteType) {
+        this.openNewRoute = openNewRoute;
         this.getIncomes().then();
     }
 
-    async getIncomes() {
-        const result = await HttpUtils.request('/categories/income');
+    private async getIncomes(): Promise<void> { // Получаем список категорий доходов
+        const result = await HttpUtils.request<GetCategoryType[]>('/categories/income');
         if (result.redirect) { // если нужно перенаправление
             return this.openNewRoute(result.redirect);
         }
 
-        if (result.error || !result.response || (result.response && result.response.error)) {
+        if (result.error || !result.response || (result.response && result.error)) {
             return alert('Возникла ошибка при получении доходов. Обратитесь в поддержку.');
         }
 
         this.showIncomes(result.response);
     }
 
-    showIncomes(incomes) {
-        const incomesListElement = document.getElementById('incomes-list');
+    private showIncomes(incomes: GetCategoryType[]): void {
+        const incomesListElement: HTMLElement | null = document.getElementById('incomes-list');
+        if (incomesListElement === null) {
+            throw new Error('Элемент для отображения списка категорий доходов не найден');
+        }
         for (let i = 0; i < incomes.length; i++) {
-            const column = document.createElement('div');
+            const column: HTMLDivElement = document.createElement('div');
             column.className = ('col-12 col-sm-6 col-lg-4');
-            const card = document.createElement('div');
+            const card: HTMLDivElement = document.createElement('div');
             card.className = ('card h-100');
-            const cardBody = document.createElement('div');
+            const cardBody: HTMLDivElement = document.createElement('div');
             cardBody.className = ('card-body');
-            const cardTitle = document.createElement('h5');
+            const cardTitle: HTMLHeadingElement = document.createElement('h5');
             cardTitle.className = ('card-title');
             cardTitle.innerText = incomes[i].title;
-            const buttonGroup = document.createElement('div');
+            const buttonGroup: HTMLDivElement = document.createElement('div');
             buttonGroup.className = ('d-grid gap-2 d-sm-flex');
-            const editButton = document.createElement('a');
+            const editButton: HTMLAnchorElement = document.createElement('a');
             editButton.className = ('btn btn-primary');
             editButton.href = '/incomes/edit?id=' + incomes[i].id;
             editButton.innerText = 'Редактировать';
-            const deleteButton = document.createElement('button');
+            const deleteButton: HTMLButtonElement = document.createElement('button');
             deleteButton.addEventListener('click', () => {
                 this.deleteCategory(incomes[i].id);
             });
             deleteButton.className = ('btn btn-danger');
-            deleteButton.setAttribute('data-id', incomes[i].id);
+            deleteButton.setAttribute('data-id', String(incomes[i].id));
             deleteButton.innerText = 'Удалить';
 
             cardBody.appendChild(cardTitle);
@@ -57,15 +63,15 @@ export class Incomes {
         }
 
 
-        const column = document.createElement('div');
+        const column: HTMLDivElement = document.createElement('div');
         column.className = ('col-12 col-sm-6 col-lg-4');
-        const card = document.createElement('div');
+        const card: HTMLDivElement = document.createElement('div');
         card.className = ('card h-100');
-        const cardBody = document.createElement('div');
+        const cardBody: HTMLDivElement = document.createElement('div');
         cardBody.className = ('card-body d-flex justify-content-center align-items-center');
-        const addButton = document.createElement('a');
+        const addButton: HTMLAnchorElement = document.createElement('a');
         addButton.href = '/incomes/create';
-        const svgIcon = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
+        const svgIcon: SVGSVGElement = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
         svgIcon.setAttribute('width', '16');
         svgIcon.setAttribute('height', '33');
         svgIcon.setAttribute('fill', '#CED4DA');
@@ -79,18 +85,24 @@ export class Incomes {
         incomesListElement.appendChild(column);
     }
 
-    deleteCategory(id) {
-        const modalElement = document.getElementById('deleteModal');
+    private deleteCategory(id: number): void {
+        const modalElement: HTMLElement | null = document.getElementById('deleteModal');
+        if (modalElement === null) {
+            throw new Error('Модальное окно для подтверждения удаления не найдено');
+        }
         const modalInstance = new bootstrap.Modal(modalElement);
         modalInstance.show();
-        const deleteButton = document.getElementById('deleteButton');
+        const deleteButton: HTMLElement | null = document.getElementById('deleteButton');
+        if (deleteButton === null) {
+            throw new Error('Кнопка подтверждения удаления не найдена');
+        }
         deleteButton.onclick = () => {
             modalInstance.hide();
             this.deleteIncome(id).then();
         };
     }
 
-    async deleteIncome(id) {
+    private async deleteIncome(id: number): Promise<void> {
         const result = await HttpUtils.request('/categories/income/' + id, 'DELETE', true);
         if (result.redirect) {
             return this.openNewRoute(result.redirect);
@@ -102,5 +114,4 @@ export class Incomes {
 
         return this.openNewRoute('/incomes');
     }
-
 }
